refactor: move file upload helpers to separate file
